+++ lib/src/objects/batch.dart
@@ -4,14 +4,27 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 
 class SyncBatch {
   final WriteBatch batch;
-  final Completer<void> completer;
+  final List<Future Function()> asyncCallbacks;
+  final List<Function()> syncCallbacks;
 
   SyncBatch()
       : batch = FirebaseFirestore.instance.batch(),
-        completer = Completer<void>();
+        asyncCallbacks = [],
+        syncCallbacks = [];
 
   Future<void> commit() async {
     await batch.commit();
-    completer.complete();
+    await Future.wait(asyncCallbacks.map((callback) => callback()));
+    for (final callback in syncCallbacks) {
+      callback();
+    }
+  }
+
+  void addAsyncCallback(Future Function() callback) {
+    asyncCallbacks.add(callback);
+  }
+
+  void addSyncCallback(Function() callback) {
+    syncCallbacks.add(callback);
   }
 }